# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillSwap Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface SignupInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  bio?: string;
  location?: string;
  skillsOffered?: string[];
  skillsWanted?: string[];
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  bio?: string;
  location?: string;
  profileImage?: string;
  skillsOffered: string[];
  skillsWanted: string[];
  isBlocked: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserWithMatch = User & {
  matchingSkills: string[];
  matchScore: number;
};

export interface UpdateProfileInput {
  name?: string;
  bio?: string;
  location?: string;
  profileImage?: string;
  skillsOffered?: string[];
  skillsWanted?: string[];
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminUserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export type SwapRequestStatus =
  (typeof SwapRequestStatus)[keyof typeof SwapRequestStatus];

export const SwapRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface SwapRequest {
  id: number;
  senderId: number;
  receiverId: number;
  skillOffered: string;
  skillRequested: string;
  message?: string;
  status: SwapRequestStatus;
  sender?: User;
  receiver?: User;
  createdAt: string;
  updatedAt?: string;
}

export interface CreateRequestInput {
  receiverId: number;
  skillOffered: string;
  skillRequested: string;
  message?: string;
}

export interface Notification {
  id: number;
  userId: number;
  message: string;
  isRead: boolean;
  type?: string;
  relatedId?: number;
  createdAt: string;
}

export interface PlatformStats {
  totalUsers: number;
  totalRequests: number;
  completedSwaps: number;
  activeUsers: number;
  pendingRequests: number;
  acceptedRequests: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  userName?: string;
  createdAt: string;
}

export interface SkillCount {
  skill: string;
  count: number;
}

export type GetUsersParams = {
  search?: string;
  skill?: string;
  page?: number;
  limit?: number;
};

export type GetRequestsParams = {
  type?: GetRequestsType;
  status?: GetRequestsStatus;
};

export type GetRequestsType =
  (typeof GetRequestsType)[keyof typeof GetRequestsType];

export const GetRequestsType = {
  sent: "sent",
  received: "received",
  all: "all",
} as const;

export type GetRequestsStatus =
  (typeof GetRequestsStatus)[keyof typeof GetRequestsStatus];

export const GetRequestsStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export type AdminGetUsersParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type AdminGetRequestsParams = {
  status?: string;
  page?: number;
};