# lib/api-zod/src/generated/types/swapRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillSwap Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SwapRequestStatus } from "./swapRequestStatus";
import type { User } from "./user";

export interface SwapRequest {
  id: number;
  senderId: number;
  receiverId: number;
  skillOffered: string;
  skillRequested: string;
  message?: string;
  status: SwapRequestStatus;
  sender?: User;
  receiver?: User;
  createdAt: Date;
  updatedAt?: Date;
}