# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillSwap Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  bio?: string;
  location?: string;
  profileImage?: string;
  skillsOffered: string[];
  skillsWanted: string[];
  isBlocked: boolean;
  createdAt: Date;
}